refactor: Sidebar.tsx를 Header, Nav, Profile 등 기능별로 분할(#48)
